New types uword_t, sword_t
First commit in a series of changes for LLP64 platform support.

We have been using `(un)signed long' assuming that it is an integer
of the same size as a pointer.  On Windows x64 this assumption does
not hold, and `(un)signed long long' is required for the same
purpose.

Code will be updated to use our own types uword_t, sword_t instead,
which are defined to be synonyms of uintptr_t and intptr_t, respectively.

Note that uword_t is also synonymous with our older type
pointer_sized_uint_t, and technically (albeit not conceptually) the
same as lispobj.

Thanks to Anton Kovalenko.

+++ src/runtime/gc.h
@@ -17,10 +17,11 @@
 #define _GC_H_
 
 #include "sbcl.h"
+#include <stdint.h>
 
 #define PAGE_BYTES BACKEND_PAGE_BYTES
 
-typedef signed long page_index_t;
+typedef intptr_t page_index_t;
 #define PAGE_INDEX_FMT "ld"
 
 typedef signed char generation_index_t;
